Add board.rb file
