+++ bin/board.rb
@@ -1,0 +1,12 @@
+class Board
+  
+  def display_board(board)
+    puts "\n #{board[0]} | #{board[1]} | #{board[2]} "
+    puts '-----------'
+    puts " #{board[3]} | #{board[4]} | #{board[5]} "
+    puts '-----------'
+    puts " #{board[6]} | #{board[7]} | #{board[8]} "
+  end
+
+end
+
