+++ bin/main.rb
@@ -1,24 +1,7 @@
 #!/usr/bin/env ruby
-# puts 'Hello World!'
 
-class TicTacToe
-    attr_accessor :broad
+require_relative "board.rb"
 
-    def initialize(board)
-        @board = ['', '', '', '', '', '', '', '', '']
-    end
-
-    WIN_COMBINATIONS = [ 
-        [0,1,2], # top_row 
-        [3,4,5], # middle_row 
-        [6,7,8], # bottom_row 
-        [0,3,6], # left_column 
-        [1,4,7], # center_column 
-        [2,5,8], # right_column 
-        [0,4,8], # left_diagonal 
-        [6,4,2] # right_diagonal 
-        ]
-    
 puts "\nWelcome to the Ruby's Tic Tac Toe game"
 
 puts "\nEnter your player 1 name"
@@ -31,15 +14,10 @@ puts "\n#{player1} will play with x and #{player2} will play with o"
 
 puts "\nLet's start the game"
 puts "\nChoose the letters between 1 and 9 to place your mark on the board"
- 
-def display_board(board)
-  puts "\n #{board[0]} | #{board[1]} | #{board[2]} "
-  puts '-----------'
-  puts " #{board[3]} | #{board[4]} | #{board[5]} "
-  puts '-----------'
-  puts " #{board[6]} | #{board[7]} | #{board[8]} "
-end
-#display_board(board)
+
+board = [1, 2, 3, 4, 5, 6, 7, 8, 9]
+show_board = Board.new 
+show_board.display_board(board)
 
 puts "\n#{player1} Enter a number between 1 and 9"
 user_input1 = gets.chomp
@@ -57,4 +35,31 @@ case user_input2.to_i
 when 1 then puts 'You are the winner!'
 when 2 then puts 'It is a tie!'
 end
-end
+
+
+
+
+
+
+
+=begin
+class TicTacToe
+    attr_accessor :board
+
+    def initialize(board)
+        @board = ['', '', '', '', '', '', '', '', '']
+    end
+
+    WIN_COMBINATIONS = [ 
+        [0,1,2], # top_row 
+        [3,4,5], # middle_row 
+        [6,7,8], # bottom_row 
+        [0,3,6], # left_column 
+        [1,4,7], # center_column 
+        [2,5,8], # right_column 
+        [0,4,8], # left_diagonal 
+        [6,4,2] # right_diagonal 
+        ]
+    
+
+=end
